relocate files to follow best practice of nextjs 13
