+++ src/lib/registry.tsx
